+++ src/Octoshift/Commands/download-logs/DownloadLogsCommandBase.cs
