+++ src/Octoshift/Commands/generate-mannequin-csv/GenerateMannequinCsvCommandBase.cs
