+++ src/Octoshift/Commands/grant-migrator-role/GrantMigratorRoleCommandBase.cs
