+++ src/Octoshift/Commands/reclaim-mannequin/ReclaimMannequinCommandBase.cs
