+++ src/Octoshift/Commands/reclaim-mannequin/ReclaimMannequinCommandHandler.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Octoshift;
 using OctoshiftCLI.Commands;
 
-namespace OctoshiftCLI.Handlers;
+namespace OctoshiftCLI.Commands;
 
 public class ReclaimMannequinCommandHandler : ICommandHandler<ReclaimMannequinCommandArgs>
 {
