+++ src/Octoshift/Commands/revoke-migrator-role/RevokeMigratorRoleCommandBase.cs
