+++ src/Octoshift/Commands/wait-for-migration/WaitForMigrationCommandBase.cs
