+++ src/OctoshiftCLI.Tests/Octoshift/Handlers/CreateTeamCommandHandlerTests.cs
@@ -2,7 +2,6 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Moq;
 using OctoshiftCLI.Commands;
-using OctoshiftCLI.Handlers;
 using Xunit;
 
 namespace OctoshiftCLI.Tests.Octoshift.Commands;
